Enables queries from BioMole to neo4j
Data is still currently set to hierachy.json and strategies.json. Neo4j data currently being used does not have enough information with it to be visualized by BioMole. When usuable data is stored in ane4j database, the code must be edited to format the data retrieved by search.py into a format similar to hierarchy.json and strategies.json.

petal-site folder was changed to petal_site so that it is readable by python.
search.py, contractions.py, and cleaner.py was moved from the bird folder to petal_site folder so that it is accessible by both bird, biomole and eventually vision.